lessons_2_3/api/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.1
black==22.12.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.2
django-filter==22.1
django-taggit==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docformatter==1.5.1
flake8==6.0.0
idna==3.4
iniconfig==1.1.1
isort==5.11.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
unify==0.5
untokenize==0.1.1
uritemplate==4.1.1
urllib3==1.26.12
